refactor nextflowhandler
