+++ cg/cli/workflow/rnafusion/base.py
@@ -31,11 +31,11 @@ from cg.constants.constants import DRY_RUN, CaseActions, MetaApis
 from cg.constants.tb import AnalysisStatus
 from cg.exc import CgError, DecompressionNeededError, MetricsQCError
 from cg.meta.workflow.analysis import AnalysisAPI
-from cg.meta.workflow.nextflow_common import NextflowAnalysisAPI
 from cg.meta.workflow.rnafusion import RnafusionAnalysisAPI
 from cg.models.cg_config import CGConfig
 from cg.models.rnafusion.command_args import CommandArgs
 from cg.store import Store
+from cg.utils.nf_handlers import NextflowHandler
 
 LOG = logging.getLogger(__name__)
 
@@ -114,21 +114,21 @@ def run(
 
     command_args: CommandArgs = CommandArgs(
         **{
-            "log": NextflowAnalysisAPI.get_log_path(
+            "log": NextflowHandler.get_log_path(
                 case_id=case_id,
                 pipeline=analysis_api.pipeline,
                 root_dir=analysis_api.root_dir,
                 log=log,
             ),
-            "work_dir": NextflowAnalysisAPI.get_workdir_path(
+            "work_dir": NextflowHandler.get_workdir_path(
                 case_id=case_id, root_dir=analysis_api.root_dir, work_dir=work_dir
             ),
             "resume": not from_start,
             "profile": analysis_api.get_profile(profile=profile),
             "with_tower": with_tower,
             "stub": stub,
-            "config": NextflowAnalysisAPI.get_nextflow_config_path(nextflow_config=config),
-            "params_file": NextflowAnalysisAPI.get_params_file_path(
+            "config": NextflowHandler.get_nextflow_config_path(nextflow_config=config),
+            "params_file": NextflowHandler.get_params_file_path(
                 case_id=case_id, root_dir=analysis_api.root_dir, params_file=params_file
             ),
             "name": case_id,
